update v2.2
alert & form validation feature

+++ application/controllers/Dashboard.php
@@ -14,6 +14,7 @@ class dashboard extends CI_Controller {
         $this->load->model('mUser');
         $this->load->model('mTransaksi');
         $this->load->model('mAkun');
+        $this->load->library("form_validation");        
         
     }
     
@@ -36,65 +37,79 @@ class dashboard extends CI_Controller {
 
     public function akun()
     {
+
         $data['allAkun'] = $this->mAkun->getAllAkun();
-        $this->load->view('akun',$data);
+        
+        $this->form_validation->set_rules('name', 'Name', 'required');
+        $this->form_validation->set_rules('noreg', 'No Rekening', 'required');
+        $this->form_validation->set_rules('alamat', 'Alamat', 'required');
+
+        if($this->form_validation->run() == false){
+            $this->load->view('akun',$data);  
+        }else{
+            $nama = $this->input->post('name');
+            $noreg = $this->input->post('noreg');
+            $alamat = $this->input->post('alamat');
+    
+            $data = array(
+                'id_akun' => $noreg,
+                'nama' => "$nama",
+                'alamat' => "$alamat"
+                 
+            );
+    
+            $this->db->insert('akun',$data);
+            $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Data berhasil disimpan</div>');
+            redirect(base_url('dashboard/akun'));
+        }
     }
 
-    public function akunInsert()
-    {
-        $nama = $this->input->post('name');
-        $noreg = $this->input->post('noreg');
-        $alamat = $this->input->post('alamat');
-
-        $data = array(
-            'id_akun' => $noreg,
-            'nama' => "$nama",
-            'alamat' => "$alamat"
-        );
-
-        $this->db->insert('akun',$data);
-        $this->session->set_flashdata('simpan','<div class="alert alert-success" role="alert">Data telah disimpan!</div>');
-        redirect(base_url('dashboard/akun'));
-    }
 
     public function transaksi()
     {
         $data['allAkun'] = $this->mAkun->getAllAkun();
         $data['allTransaksi'] = $this->mTransaksi->getAllTransaksi();
 
-        $this->load->view('transaksi',$data);
-    }
-    public function transaksiInsert(){
-        $noreg = $this->input->post('noreg');
-        $metode = $this->input->post('metode');
-        $nominal = $this->input->post('nominal');
-        $waktu = date('Y-m-d');
-        
-        if($metode == "debet"){
-            $this->db->select_sum('transaksi');
-            $this->db->where('id_akun',$noreg);
-            $query = $this->db->get('transaksi')->row_array();            
-            if((int)$query['transaksi'] <= $nominal){
-                $this->session->set_flashdata('pesan','<div class="alert alert-danger" role="alert">Data gagal disimpan!</div>');
-                redirect(base_url('dashboard/transaksi'));                
+       
+
+        $this->form_validation->set_rules('metode', 'Metode', 'required');
+        $this->form_validation->set_rules('noreg', 'No Rekening', 'required');
+        $this->form_validation->set_rules('nominal', 'Nominal', 'required');
+
+        if($this->form_validation->run() == false){
+            $this->load->view('transaksi',$data);
+        }else{
+            $noreg = $this->input->post('noreg');
+            $metode = $this->input->post('metode');
+            $nominal = $this->input->post('nominal');
+            $waktu = date('Y-m-d');
+            
+            if($metode == "debet"){
+                $this->db->select_sum('transaksi');
+                $this->db->where('id_akun',$noreg);
+                $query = $this->db->get('transaksi')->row_array();            
+                if((int)$query['transaksi'] <= $nominal){                
+                    redirect(base_url('dashboard/transaksi'));                
+                }else{
+                    $data = array(
+                        'id_akun' => $noreg,
+                        'tanggal' => $waktu,
+                        'transaksi' => -$nominal
+                    );
+                }
             }else{
                 $data = array(
                     'id_akun' => $noreg,
                     'tanggal' => $waktu,
-                    'transaksi' => -$nominal
+                    'transaksi' => (int)$nominal
                 );
             }
-        }else{
-            $data = array(
-                'id_akun' => $noreg,
-                'tanggal' => $waktu,
-                'transaksi' => (int)$nominal
-            );
+    
+            $this->db->insert('transaksi',$data);
+            $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Transaksi berhasil</div>');
+            redirect(base_url('dashboard/transaksi'));
         }
 
-        $this->db->insert('transaksi',$data);
-        $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Data telah disimpan!</div>');
-        redirect(base_url('dashboard/transaksi'));
     }
 
     public function history()
@@ -106,36 +121,39 @@ class dashboard extends CI_Controller {
     {
         $data['allAkun'] = $this->mAkun->getAllAkun();
 
-        $this->load->view('masterakun',$data);
-    }
-    public function masterAkunProses()
-    {
-        $noreg = $this->input->post('noreg');
-        $nama = $this->input->post('nama');
-
-        $simpan = $this->input->post('simpan');
-        $hapus = $this->input->post('hapus');
-
-        if(isset($simpan))
-        {
-            $this->db->set('nama',$nama);
-            $this->db->where('id_akun',$noreg);
-            $this->db->update('akun');
-            $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Data telah disimpan!</div>');          
-        }
+            
+        $this->form_validation->set_rules('noreg', 'No Rekening', 'required');
         
-        if(isset($hapus))
-        {            
+        if($this->form_validation->run() == false){
+            $this->load->view('masterakun',$data);
+        }else{
+            $noreg = $this->input->post('noreg');
+            $nama = $this->input->post('nama');
+    
+            $simpan = $this->input->post('simpan');
+            $hapus = $this->input->post('hapus');
+    
+            if(isset($simpan))
+            {
+                $this->db->set('nama',$nama);
+                $this->db->where('id_akun',$noreg);
+                $this->db->update('akun');
+                $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Data berhasil disimpan</div>');
+            }
+            
+            if(isset($hapus))
+            {            
+    
+                $this->db->where('id_akun',$noreg);
+                $this->db->delete('transaksi');
+                $this->db->where('id_akun',$noreg);
+                $this->db->delete('akun');
+                $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Data berhasil dihapus</div>');
 
-            $this->db->where('id_akun',$noreg);
-            $this->db->delete('transaksi');
-            $this->db->where('id_akun',$noreg);
-            $this->db->delete('akun');
-            $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Data telah dihapus!</div>');          
-
+            }
+            redirect(base_url('dashboard/masterakun'));
         }
-        
-        redirect(base_url('dashboard/masterakun'));
+
     }
 
     public function masteruser()
@@ -150,7 +168,7 @@ class dashboard extends CI_Controller {
         $this->db->set('level',1);
         $this->db->where('username',$param);
         $this->db->update('user');
-        $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Akun telah disimpan!</div>');          
+        $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Akun berhasil diregistrasi</div>');
         redirect(base_url('dashboard/masteruser'));
     }
 
@@ -158,7 +176,7 @@ class dashboard extends CI_Controller {
     {        
         $this->db->where('username',$param);
         $this->db->delete('user');
-        $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Data telah dihapus!</div>');          
+        $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Akun berhasil dihapus</div>');
         redirect(base_url('dashboard/masteruser'));
     }
 
@@ -170,33 +188,39 @@ class dashboard extends CI_Controller {
     }
     public function change()
     {
-        $this->load->view('change');
-    }
-    public function changeProses()
-    {
-        $oldpass = $this->input->post('oldpass');
-        $newpass = $this->input->post('newpass');        
-        $newpass2 = $this->input->post('newpass2');
 
+        $this->form_validation->set_rules('oldpass', "Password Lama", 'required');
+        $this->form_validation->set_rules('newpass', "Password Baru", 'required');
+        $this->form_validation->set_rules('newpass2', "Konfirmasi Password Baru", 'required');
         
-        if($newpass == $newpass2){
-            $this->db->set('password',$newpass);
-            $this->db->where('password',$oldpass);
-            $this->db->where('username',$this->session->nama);        
-            $this->db->update('user');
-
-            $this->session->sess_destroy();        
-            $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Password sudah diganti</div>');              
-            redirect(base_url('dashboard/'));
+        if($this->form_validation->run() == false){
+            $this->load->view('change');            
         }else{
-            $this->session->set_flashdata('pesan','<div class="alert alert-danger" role="alert">Gagal mengubah!</div>');          
-            redirect(base_url('dashboard/change'));
+            $oldpass = $this->input->post('oldpass');
+            $newpass = $this->input->post('newpass');        
+            $newpass2 = $this->input->post('newpass2');
+    
+            
+            if($newpass == $newpass2){
+                $this->db->set('password',$newpass);
+                $this->db->where('password',$oldpass);
+                $this->db->where('username',$this->session->nama);        
+                $this->db->update('user');
+    
+                $this->session->sess_destroy();   
+                $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Password berhasil diubah</div>');
+                redirect(base_url('dashboard/'));
+            }else{
+                $this->session->set_flashdata('pesan', '<div class="alert alert-danger" role="alert">Gagal ganti password</div>');
+                redirect(base_url('dashboard/change'));
+            }
         }
+
     }
     public function Logout()
     {
         $this->session->sess_destroy();
-        $this->session->set_flashdata('pesan','<div class="alert alert-success" role="alert">Anda telah logout!</div>');          
-        redirect('login');
+        $this->session->set_flashdata('pesan', '<div class="alert alert-success" role="alert">Berhasil Logout</div>');
+        redirect('dashboard');
     }
 }
